Make separate commands completer
