+++ pupy/commands/__init__.py
@@ -6,20 +6,7 @@ import os
 import imp
 import shlex
 
-def commands_completer(module, args, text, context):
-    aliases = dict(context.config.items('aliases'))
-    modules = list(context.server.iter_modules(
-        by_clients=True,
-        clients_filter=context.handler.default_filter))
-    commands = context.commands.list(False)
-
-    return [
-        x+' ' for x in aliases.iterkeys() if x.startswith(text)
-    ] + [
-        x+' ' for x,_ in commands if x.startswith(text)
-    ] + [
-        x.get_name()+' ' for x in modules if x.get_name().startswith(text)
-    ]
+from pupylib.PupyCompleter import commands_completer
 
 class InvalidCommand(Exception):
     pass
